requirements.txt
# AI Vision Defect Detection Platform
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
tqdm>=4.65.0
pyyaml>=6.0
